nifi/requirements.txt
attrs==19.1.0
great-expectations==0.5.1
pyrsistent==0.14.11
pytz==2018.9
PyYAML==3.13
mysql-connector==2.2.9
jsonschema==3.0.1
numpy==1.16.2
pandas==0.24.1
phonenumbers==8.10.7
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
xlrd==1.2.0

